refactor: extract useToastsManager
